Title: 쪼개어 배낭 채우기 구현, Time: 972ms, Memory: 52MB, Status: Passed - CodeTree

+++ 231225/쪼개어 배낭 채우기 구현/implement-fractional-knapsack.java
@@ -39,7 +39,7 @@ public class Main {
             int w = sc.nextInt();
             int v = sc.nextInt();
 
-            nap[i] = new Knapsack(w, v, v / w);
+            nap[i] = new Knapsack(w, v, (double) v / w);
         }
 
         Arrays.sort(nap, 0, n);
